Change the port type to int32_t

+++ rfaas/include/rfaas/resources.hpp
@@ -15,12 +15,12 @@ namespace rfaas {
 
   struct server_data
   {
-    int16_t port;
+    int32_t port;
     int16_t cores;
     char address[16];
 
     server_data();
-    server_data(const std::string & ip, int16_t port, int16_t cores);
+    server_data(const std::string & ip, int32_t port, int16_t cores);
 
     template <class Archive>
     void save(Archive & ar) const
